re-organize components, turn off 'apps'
