+++ scripts/generate-app-config/generateAppConfig.js
@@ -118,7 +118,7 @@ const generateAppConfig = (siteConfig, production, verbose) => {
     contentConfig: contentConfigImport,
     navigationItems: navigationItemsImport,
     indexPath: navConfig.navigation.index,
-    apps: siteConfig.apps,
+    apps: [], //siteConfig.apps,
     capabilities: capabilities.config,
     extensions: extensionConfigImport,
     placeholderSrc: imports.addImport(placeholderSrc, 'placeholderSrc'),
